Don't change zoom level when marking a store visited/unvisited
Currently, when marking a store visited/unvisited,
UI is fully rerendered. This causes a zoom level change
when a search term is set, because UI would be updated
based on the search term.
This commit fixes it by not fully updating the UI when a store
is marked visited/unvisited.