Fix error for "show on map"
"Show on map" sometimes does not work because map rerendering
is asynchronous and it overwrites the behavior to zoom into
a single store.
While zooming into a store was used with a delay (settimeout) and
increasing the delay value would solve the problem,
it may be underterministic.

Instead, solve this problem by passing the store information
to the rerendering function and let it zoom in.